nested: Fix 64-bit compilation

+++ nested.d
@@ -27,7 +27,7 @@ class NestedFormatter
 	}
 
 	Block[] blocks;
-	int blockIndex;
+	size_t blockIndex;
 
 	this(Block[] blocks)
 	{
